feat(ImageCarousel): enhance modulation and update documentation
- Refined the modulation approach for the ImageCarousel component to improve its flexibility and integration capabilities.
- Rename ImageCarousel to CarouselDisplay
- Expanded and updated the documentation to provide clearer guidance on usage, configuration options, and examples.
chore: removed unnecessary items
