+++ public/css/ui.css
@@ -29,7 +29,7 @@ body {
     height: 100vh;
 }
 
-#Test{
+#mask{
     position: absolute;
     width: 100%;
     height: 100%;
