+++ public/index.html
@@ -9,7 +9,7 @@
         <div id="container" class="hidden">
             <div id="carousel">
                 <img id="background-img" class="background">
-                <div id="Test">
+                <div id="mask">
                 </div>
                 <div class="title-container">
                     <div>
@@ -26,7 +26,7 @@
                 </div>
             </div>
             <img id="logo" src="./images/PeacockLogo.png">
-        </img>
+        </div>
         <!-- <cast-media-player></cast-media-player> -->
         <!-- <script src="./js/Carrosel.js"></script> -->
         <script type="module" src="./js/Main.js"></script>
