+++ public/js/communication/MessageProtocol.js
@@ -10,7 +10,6 @@ class MessageProtocol {
     constructor(type = "", data = "") {
         this.type = type;
         this.data = data;
-        // this.media = data;
     }
 
     /**
